Reorganize src/ and tests/ into feature folders
Moves flat src/ into src/features/{meals,workouts}, src/integrations/google,
and src/core/{db,http,config,util}. Mirrors the layout in tests/. Includes
are rewritten to path-qualified form (e.g. "features/meals/meal.h") so the
include root remains src/ — no per-subdirectory include flags needed.

No logic changes; build + 101 tests still green.
