+++ CMakeLists.txt
@@ -70,29 +70,27 @@ find_package(SQLite3 REQUIRED)
 # --- Source Files ---
 # Explicitly listing sources is preferred over file(GLOB ...)
 set(LIB_SOURCES
-    src/db_manager.cpp
-    src/meal_factory.cpp
-    src/meal_planner.cpp
-    src/api_routes.cpp
-    src/config_parser.cpp
-    src/google_oauth.cpp
-    src/calendar_service.cpp
-    src/token_encryption.cpp
-    # Add other library source files here
+    src/core/db/db_manager.cpp
+    src/core/config/config_parser.cpp
+    src/core/http/api_routes.cpp
+    src/features/meals/meal_factory.cpp
+    src/features/meals/meal_planner.cpp
+    src/integrations/google/google_oauth.cpp
+    src/integrations/google/calendar_service.cpp
+    src/integrations/google/token_encryption.cpp
 )
 
 set(MAIN_SOURCE src/main.cpp)
 
 set(TEST_SOURCES
-    tests/test_ingredient.cpp
-    tests/test_meal.cpp
-    tests/test_meal_factory.cpp
-    tests/test_measurement.cpp
-    tests/test_meal_planner.cpp
-    tests/test_db_manager.cpp
-    tests/test_token_encryption.cpp
-    tests/test_workout_service.cpp
-    # Add other test source files here
+    tests/features/meals/test_ingredient.cpp
+    tests/features/meals/test_meal.cpp
+    tests/features/meals/test_meal_factory.cpp
+    tests/features/meals/test_measurement.cpp
+    tests/features/meals/test_meal_planner.cpp
+    tests/features/workouts/test_workout_service.cpp
+    tests/core/db/test_db_manager.cpp
+    tests/integrations/google/test_token_encryption.cpp
 )
