+++ src/core/config/config_parser.cpp
@@ -1,4 +1,4 @@
-#include "config_parser.h"
+#include "core/config/config_parser.h"
 
 #include <crow.h>
